+++ app/src/main/java/com/owncloud/android/operations/RefreshFolderOperation.java
@@ -16,8 +16,8 @@ import com.google.gson.Gson;
 import com.nextcloud.android.lib.resources.directediting.DirectEditingObtainRemoteOperation;
 import com.nextcloud.client.account.User;
 import com.nextcloud.common.NextcloudClient;
+import com.nextcloud.utils.ResultParser;
 import com.nextcloud.utils.e2ee.E2EVersionHelper;
-import com.nextcloud.utils.extensions.RemoteOperationResultExtensionsKt;
 import com.nextcloud.utils.extensions.StringExtensionsKt;
 import com.owncloud.android.datamodel.ArbitraryDataProvider;
 import com.owncloud.android.datamodel.ArbitraryDataProviderImpl;
@@ -304,7 +304,8 @@ public class RefreshFolderOperation extends RemoteOperation {
             sendLocalBroadcast(EVENT_SINGLE_FOLDER_CONTENTS_SYNCED, mLocalFolder.getRemotePath(), result);
         }
 
-        final List<RemoteFile> remoteFiles = RemoteOperationResultExtensionsKt.dataOfType(result, RemoteFile.class);
+
+        final var remoteFiles = ResultParser.dataOfType(result, RemoteFile.class);
         if (!remoteFiles.isEmpty() && !mSyncFullAccount && !mOnlyFileMetadata) {
             // this needed because if file has new share or share is removed, eTag is not changing.
             // that's why another separate EVENT_SINGLE_FOLDER_SHARES_SYNCED introduced before.
